requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
branca==0.5.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
dominate==2.6.0
entrypoints==0.4
executing==0.8.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
folium==0.12.1.post1
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
itsdangerous==2.0.1
jedi==0.18.1f
Jinja2==3.0.1
jupyter-client==7.3.4
jupyter-core==4.10.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
parso==0.8.3
peewee==3.14.10
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.17.1
pyzmq==23.1.0
requests==2.27.1
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.2.post1
urllib3==1.26.9
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1